Know what? I'll just export a Markdown version for GitHub. Also typos
I was hoping that I could just get the org file to look nice enough in
GitHub's interface, but it appears that the results are ugly unless I
make my Org file ugly.

I also forgot to update the README with my new goals, so those are in
there now.