Translated Masara town and got confused by grammar
The town is pretty small, but there are a few things that the two NPCs
say that were pretty difficult to figure out. For example, the と at
the end of the sentence spoken by techno-guy was a bit of an enigma,
and was also not easy to look up. In fact, it might still be wrong. (I
hope not though..!) The other thing that was odd was the を leading up
to a noun also in techno-guy's speech. I've only seen it used to pair
something on the left to a verb or modified verb on the right.

Wow techno-dude, you are one hard-to-read dude.